feat: move check, cleanup
